fix: up-next plant render issue
